pyproject.toml
[tool.poetry]
name = "tracecheck"
version = "0.4.0"
description = "Evaluate multi step AI agent traces, not just single LLM responses."
authors = ["Mohd Talib <mohd.talib@growthengineering.co.uk>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "tracecheck" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6"
typer = ">=0.12"
pyyaml = "^6.0"
anthropic = { version = ">=0.40", optional = true }

[tool.poetry.extras]
anthropic = ["anthropic"]
llm = ["anthropic"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.scripts]
tracecheck = "tracecheck.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
